ENG-1088 Refactor rtl test utils to have its own module
