+++ test/ui/email-config/AddEmailSenderFormContainer.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { fireEvent, screen } from '@testing-library/dom';
+import { fireEvent, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import * as reactRedux from 'react-redux';
 
-import { renderWithIntlAndState, setupForm } from 'test/testUtils';
+import { renderWithIntlAndState, setupForm } from 'test/rtlTestUtils';
 import userEvent from '@testing-library/user-event';
 import AddEmailSenderFormContainer from 'ui/email-config/AddEmailSenderFormContainer';
