+++ test/ui/email-config/EditEmailSenderFormContainer.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { fireEvent, screen } from '@testing-library/dom';
+import { fireEvent, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import * as reactRedux from 'react-redux';
 
-import { renderWithIntlRouterState, setupForm } from 'test/testUtils';
+import { renderWithIntlRouterState, setupForm } from 'test/rtlTestUtils';
 import userEvent from '@testing-library/user-event';
 import EditEmailSenderFormContainer from 'ui/email-config/EditEmailSenderFormContainer';
