+++ test/ui/email-config/EmailConfigSenderMgmtContainer.test.js
@@ -1,10 +1,10 @@
-import { screen, within } from '@testing-library/dom';
+import { screen, within } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 import * as reactRedux from 'react-redux';
 
-import { renderWithIntlRouterState } from 'test/testUtils';
+import { renderWithIntlRouterState } from 'test/rtlTestUtils';
 import EmailConfigSenderMgmtContainer from 'ui/email-config/EmailConfigSenderMgmtContainer';
 import { SET_VISIBLE_MODAL, SET_INFO } from 'state/modal/types';
 import { MODAL_ID as DELETE_SENDER_MODAL_ID } from 'ui/email-config/DeleteSenderModal';
