+++ test/ui/email-config/EmailConfigSmtpServerContainer.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { screen, within } from '@testing-library/dom';
+import { screen, within } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import userEvent from '@testing-library/user-event';
 import * as reactRedux from 'react-redux';
 
-import { renderWithIntlAndState, setupForm } from 'test/testUtils';
+import { renderWithIntlAndState, setupForm } from 'test/rtlTestUtils';
 import EmailConfigSmtpServerContainer from 'ui/email-config/EmailConfigSmtpServerContainer';
 
 jest.unmock('react-redux');
